requirements.txt
torch>=1.9.0
numpy>=1.21.0
pyyaml>=5.4.0
requests>=2.25.0
jsonlines>=2.0.0
matplotlib>=3.3.0
tqdm>=4.60.0